+++ _prepend.php
@@ -15,14 +15,15 @@
 if (!defined('DC_RC_PATH')) return;
 
 $__autoload['relatedHelpers'] = dirname(__FILE__).'/inc/related.helpers.php';
-$__autoload['rsRelated'] = dirname(__FILE__).'/inc/lib.related.php';
-$__autoload['adminPageList'] = dirname(__FILE__).'/inc/lib.related.php';
+$__autoload['adminPageList'] = dirname(__FILE__).'/inc/admin.page.list.php';
 $__autoload['widgetsRelated'] = dirname(__FILE__).'/inc/widgets.related.php';
 $__autoload['relatedAdminBehaviors'] = dirname(__FILE__).'/inc/related.admin.behaviors.php';
 $__autoload['relatedPublicBehaviors'] = dirname(__FILE__).'/inc/related.public.behaviors.php';
 $__autoload['relatedUrlHandlers'] = dirname(__FILE__).'/inc/related.url.handlers.php';
-$__autoload['rsRelatedBase'] = dirname(__FILE__).'/inc/rs.related.base.php';
 $__autoload['relatedTemplates'] = dirname(__FILE__).'/inc/related.templates.php';
+$__autoload['relatedPagesActionsPage'] = dirname(__FILE__).'/inc/related.pages.actionspage.php';
+$__autoload['rsRelated'] = dirname(__FILE__).'/inc/rs.related.php';
+$__autoload['rsRelatedBase'] = dirname(__FILE__).'/inc/rs.related.base.php';
 
 $self_ns = $core->blog->settings->addNamespace('related');
 if ($self_ns->active) {