+++ inc/panel.php
@@ -70,19 +70,95 @@ if (!empty($_POST['saveconfig'])) {
             if (is_dir($tmp_repository) && is_writable($tmp_repository)) {
                 $core->blog->settings->related->put('files_path', $tmp_repository);
                 $repository = $tmp_repository;
-                $message = __('Configuration updated.');
+                $message = dcPage::addSuccessNotice(__('Configuration updated.'));
             } else {
                 $core->error->add(__('Directory for related files repository needs to allow read and write access.'));
             }
         }
 
-        $message = __('The configuration has been updated.');
+        $message = dcPage::addSuccessNotice(__('The configuration has been updated.'));
     } catch(Exception $e) {
         $core->error->add($e->getMessage());
     }
 }
 
 if ($related_active) {
+    /* filters */
+    $user_id = !empty($_GET['user_id']) ? $_GET['user_id'] : '';
+    $status = isset($_GET['status']) ? $_GET['status'] : '';
+    $in_widget = isset($_GET['in_widget']) ? $_GET['in_widget'] : '';
+    $month = !empty($_GET['month']) ? $_GET['month'] : '';
+    $lang = !empty($_GET['lang']) ?	$_GET['lang'] : '';
+    $sortby = !empty($_GET['sortby']) ?	$_GET['sortby'] : 'post_dt';
+    $order = !empty($_GET['order']) ? $_GET['order'] : 'desc';
+
+    $show_filters = false;
+    $form_filter_title = __('Show filters and display options');
+
+        # Creating filter combo boxes
+    if (!$core->error->flag()) {
+        # Getting authors
+        try {
+            $users = $core->blog->getPostsUsers();
+        } catch (Exception $e) {
+            $core->error->add($e->getMessage());
+        }
+
+        # Filter form we'll put in html_block
+        $users_combo = array_merge(
+            array('-' => ''),
+            dcAdminCombos::getUsersCombo($users)
+        );
+
+        # Getting langs
+        try {
+            $langs = $core->blog->getLangs();
+        } catch (Exception $e) {
+            $core->error->add($e->getMessage());
+        }
+        $lang_combo = array_merge(
+            array('-' => ''),
+            dcAdminCombos::getLangsCombo($langs,false)
+        );
+
+        $status_combo = array_merge(
+            array('-' => ''),
+            dcAdminCombos::getPostStatusesCombo()
+        );
+
+        $in_widget_combo = array(
+            '-' => '',
+            __('yes') => 1,
+            __('no') => 0
+        );
+
+        # Getting dates
+        try {
+            $dates = $core->blog->getDates(array('type'=>'month'));
+        } catch (Exception $e) {
+            $core->error->add($e->getMessage());
+        }
+
+        # Months array
+        $dt_m_combo = array_merge(
+            array('-' => ''),
+            dcAdminCombos::getDatesCombo($dates)
+        );
+
+        $sortby_combo = array(
+            __('Date') => 'post_dt',
+            __('Title') => 'post_title',
+            __('Author') => 'user_id',
+            __('Status') => 'post_status',
+            __('Visible pages in widget') => 'post_selected'
+        );
+
+        $order_combo = array(
+            __('Descending') => 'desc',
+            __('Ascending') => 'asc'
+        );
+    }
+
     $page = !empty($_GET['page']) ? $_GET['page'] : 1;
     $nb_per_page =  30;
     if (!empty($_GET['nb']) && (integer) $_GET['nb'] > 0) {
@@ -91,6 +167,7 @@ if ($related_active) {
 
     $params['limit'] = array((($page-1)*$nb_per_page),$nb_per_page);
     $params['no_content'] = true;
+    $params['order'] = 'post_position asc';
 
     # Get pages
     try {
@@ -98,6 +175,74 @@ if ($related_active) {
         $pages->extend("rsRelated");
         $counter = $core->blog->getPosts($params,true);
         $page_list = new adminPageList($core,$pages,$counter->f(0));
+    } catch (Exception $e) {
+        $core->error->add($e->getMessage());
+    }
+
+        # apply filters
+    # - User filter
+    if ($user_id !== '' && in_array($user_id,$users_combo)) {
+        $params['user_id'] = $user_id;
+        $show_filters = true;
+    } else {
+        $user_id='';
+    }
+
+    # - Status filter
+    if ($status !== '' && in_array($status, $status_combo)) {
+        $params['post_status'] = $status;
+        $show_filters = true;
+    } else {
+        $status='';
+    }
+
+    # - in widget filter
+    if ($in_widget !== '' && in_array($in_widget, $in_widget_combo)) {
+        $params['post_selected'] = $in_widget;
+        $show_filters = true;
+    } else {
+        $in_widget = '';
+    }
+
+    # - Month filter
+    if ($month !== '' && in_array($month,$dt_m_combo)) {
+        $params['post_month'] = substr($month,4,2);
+        $params['post_year'] = substr($month,0,4);
+        $show_filters = true;
+    } else {
+        $month='';
+    }
+
+    # - Lang filter
+    if ($lang !== '' && in_array($lang,$lang_combo)) {
+        $params['post_lang'] = $lang;
+        $show_filters = true;
+    } else {
+        $lang='';
+    }
+
+    # - Sortby and order filter
+    if ($sortby !== '' && in_array($sortby,$sortby_combo)) {
+        if ($order !== '' && in_array($order,$order_combo)) {
+            $params['order'] = $sortby.' '.$order;
+        } else {
+            $order='desc';
+        }
+
+        if ($sortby != 'post_dt' || $order != 'desc') {
+            $show_filters = true;
+        }
+    } else {
+        $sortby = 'post_dt';
+        $order = 'desc';
+    }
+
+    # Get pages
+    try {
+        $pages = $core->blog->getPosts($params);
+        $pages->extend('rsRelated');
+        $counter = $core->blog->getPosts($params, true);
+        $page_list = new adminPageList($core, $pages, $counter->f(0));
     } catch (Exception $e) {
         $core->error->add($e->getMessage());
     }
@@ -115,13 +260,27 @@ if ($related_active) {
     if ($core->auth->check('delete,contentadmin',$core->blog->id)) {
         $combo_action[__('delete')] = 'delete';
     }
+    $combo_action[__('Widget')] = array(
+        __('Add to widget') => 'selected',
+        __('Remove from widget') => 'unselected'
+    );
 
     # --BEHAVIOR-- adminPagesActionsCombo
     $core->callBehavior('adminPagesActionsCombo',array(&$combo_action));
 
+    $pages_actions_page = new relatedPagesActionsPage($core, 'plugin.php', array('p' => 'related'));
+    if (!$pages_actions_page->process()) {
+        $process_successfull = false;
+    } else {
+        $process_successfull = true;
+    }
+
     /**
      * Manage public list if requested
      */
+    $all_pages = $core->blog->getPosts(array('post_type' => 'related'));
+    $all_pages->extend('rsRelated');
+
     if (isset($_POST['pages_upd'])) {
         $default_tab = 'pages_order';
 
@@ -153,9 +312,11 @@ if ($related_active) {
     }
 
     if (!empty($_GET['reord'])) {
-        $message = __('Pages list has been sorted.');
+        $message = dcPage::addSuccessNotice(__('Pages list has been sorted.'));
         $default_tab = 'pages_order';
     }
 }
 
-include_once(dirname(__FILE__).'/../tpl/index.tpl');
+if (!$process_successfull) {
+    include_once(dirname(__FILE__).'/../tpl/index.tpl');
+}